fix(wp): split mocha test from outputScripts impl
Otherwise it cannot be imported from `main.ts`

Issue: BG-34148
